Add all_dif generator

+++ src/first_alg_with_heuristics.h
@@ -9,8 +9,8 @@
 #include <cassert>
 #include <algorithm>
 
-void HammingDistanceBase(size_t n, size_t m, size_t sigma, double eps, const std::vector<uint32_t>& A, const std::vector<uint32_t>& B, std::vector<uint32_t>& dist, std::mt19937& rng);
+void HammingDistanceBase(int n, int m, int sigma, double eps, const std::vector<uint32_t>& A, const std::vector<uint32_t>& B, std::vector<uint32_t>& dist, std::mt19937& rng);
 
-void HammingDistanceHeuristic_1(size_t n, size_t m, size_t sigma, double eps, const std::vector<uint32_t>& A, const std::vector<uint32_t>& B, std::vector<uint32_t>& dist, std::mt19937& rng);
+void HammingDistanceHeuristic_1(int n, int m, int sigma, double eps, const std::vector<uint32_t>& A, const std::vector<uint32_t>& B, std::vector<uint32_t>& dist, std::mt19937& rng);
 
-void HammingDistanceHeuristic_2(size_t n, size_t m, size_t sigma, double eps, const std::vector<uint32_t>& A, const std::vector<uint32_t>& B, std::vector<uint32_t>& dist, std::mt19937& rng);
+void HammingDistanceHeuristic_2(int n, int m, int sigma, double eps, const std::vector<uint32_t>& A, const std::vector<uint32_t>& B, std::vector<uint32_t>& dist, std::mt19937& rng);
